my-cdk-project/requirements-dev.txt
pytest==6.2.5
aws-cdk-lib>=2.80.0
constructs>=10.0.0
boto3